Device cycling sensor: fixed timestamp overflow

+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/service/devices/cycling_sensor/support/CyclingSensorSupport.java
@@ -229,7 +229,7 @@ public class CyclingSensorSupport extends CyclingSensorBaseSupport {
             sample.setDistance(currentMeasurement.revolutionCount * wheelCircumference);
         }
 
-        sample.setTimestamp((int)(now));
+        sample.setTimestamp(now);
         sample.setDevice(databaseDevice);
         sample.setUser(databaseUser);
